refactor(webapp): reclassify patterns/ + move ConditionalTooltip, EditableInput, SecretInput
Redefines the ui/ vs patterns/ boundary:
- ui/      = design-system lift candidates (Storybook required)
- patterns/ = stays in webapp (Nango-specific OR generic helpers that
              don't belong in a published component library)

Moves ConditionalTooltip (conditional wrapper with no publishable visual
story), EditableInput and SecretInput (both depend on PermissionGate, a
Nango-specific pattern) from ui/ to patterns/.

Updates all import sites and AGENTS.md. Notion taxonomy updated to match.
