+++ packages/webapp/src/pages/Environment/Settings/Deprecated.tsx
@@ -11,9 +11,9 @@ import { useToast } from '../../../hooks/useToast';
 import { useStore } from '../../../store';
 import { PermissionGate } from '@/components-v2/patterns/PermissionGate';
 import { ButtonLink } from '@/components-v2/ui/Button';
-import { EditableInput } from '@/components-v2/ui/EditableInput';
+import { EditableInput } from '@/components-v2/patterns/EditableInput';
 import { Label } from '@/components-v2/ui/Label';
-import { SecretInput } from '@/components-v2/ui/SecretInput';
+import { SecretInput } from '@/components-v2/patterns/SecretInput';
 import { Switch } from '@/components-v2/ui/Switch';
 import { usePermissions } from '@/hooks/usePermissions';
