+++ packages/webapp/src/pages/Environment/Settings/General.tsx
@@ -13,8 +13,8 @@ import { useStore } from '../../../store';
 import { PermissionGate } from '@/components-v2/patterns/PermissionGate';
 import { Alert, AlertDescription } from '@/components-v2/ui/Alert';
 import { ButtonLink } from '@/components-v2/ui/Button';
-import { ConditionalTooltip } from '@/components-v2/ui/ConditionalTooltip';
-import { EditableInput } from '@/components-v2/ui/EditableInput';
+import { ConditionalTooltip } from '@/components-v2/patterns/ConditionalTooltip';
+import { EditableInput } from '@/components-v2/patterns/EditableInput';
 import { Switch } from '@/components-v2/ui/Switch';
 import { useConfirmDialog } from '@/hooks/useConfirmDialog';
 import { usePermissions } from '@/hooks/usePermissions';
