+++ packages/webapp/src/pages/Environment/Settings/Webhooks.tsx
@@ -7,9 +7,9 @@ import SettingsGroup from './components/SettingsGroup.js';
 import { WebhookCheckboxes } from './components/WebhookCheckboxes.js';
 import { useEnvironment, usePatchWebhook } from '../../../hooks/useEnvironment.js';
 import { ButtonLink } from '@/components-v2/ui/Button.js';
-import { EditableInput } from '@/components-v2/ui/EditableInput.js';
+import { EditableInput } from '@/components-v2/patterns/EditableInput.js';
 import { Label } from '@/components-v2/ui/Label.js';
-import { SecretInput } from '@/components-v2/ui/SecretInput.js';
+import { SecretInput } from '@/components-v2/patterns/SecretInput.js';
 import { usePermissions } from '@/hooks/usePermissions.js';
 import { useToast } from '@/hooks/useToast.js';
 import { validateUrl } from '@/pages/Integrations/utils.js';
