+++ packages/webapp/src/pages/Environment/Settings/components/DeleteButton.tsx
@@ -4,7 +4,7 @@ import { permissions } from '@nangohq/authz';
 
 import { PermissionGate } from '@/components-v2/patterns/PermissionGate';
 import { Button } from '@/components-v2/ui/Button';
-import { ConditionalTooltip } from '@/components-v2/ui/ConditionalTooltip';
+import { ConditionalTooltip } from '@/components-v2/patterns/ConditionalTooltip';
 import { DestructiveActionModal } from '@/components-v2/ui/DestructiveActionModal';
 import { useEnvironment } from '@/hooks/useEnvironment';
 import { usePermissions } from '@/hooks/usePermissions';
