+++ packages/webapp/src/pages/Integrations/components/forms/InstallPluginAuthCreateForm.tsx
@@ -6,7 +6,7 @@ import z from 'zod';
 import { Button } from '@/components-v2/ui/Button';
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components-v2/ui/Form';
 import { InputGroup, InputGroupInput } from '@/components-v2/ui/InputGroup';
-import { SecretInput } from '@/components-v2/ui/SecretInput';
+import { SecretInput } from '@/components-v2/patterns/SecretInput';
 
 import type { ApiProviderListItem, PostIntegration } from '@nangohq/types';
