+++ packages/webapp/src/pages/Integrations/components/forms/McpOAuthCreateForm.tsx
@@ -7,7 +7,7 @@ import { ScopesInput } from '@/components-v2/patterns/ScopesInput';
 import { Button } from '@/components-v2/ui/Button';
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components-v2/ui/Form';
 import { InputGroupInput } from '@/components-v2/ui/InputGroup';
-import { SecretInput } from '@/components-v2/ui/SecretInput';
+import { SecretInput } from '@/components-v2/patterns/SecretInput';
 
 import type { ApiProviderListItem, PostIntegration } from '@nangohq/types';
